Add MCP demo image to README and update Google Tasks MCP server
