+++ mcp_servers/google_tasks/requirements.txt
@@ -1,9 +1,9 @@
-mcp>=1.12.0
-google-api-python-client>=2.117.0
-google-auth>=2.29.0
-python-dotenv>=1.0.1
-starlette>=0.36.3
-uvicorn[standard]>=0.29.0
-click>=8.1.7
-tenacity>=8.2.3
-
+mcp
+google-api-python-client
+google-auth
+python-dotenv
+starlette
+uvicorn[standard]
+click
+tenacity
+anyio
