+++ mcp_servers/google_tasks/tools/__init__.py
@@ -1,17 +1,3 @@
-from __future__ import annotations
-
-import contextvars
-
-# Support running both as a package (python -m mcp_servers.google_tasks.server)
-# and as a script (python mcp_servers/google_tasks/server.py)
-try:  # package context
-    from ..auth import get_tasks_service  # type: ignore
-except Exception:
-    try:  # script context (auth.py is alongside this package directory)
-        from auth import get_tasks_service  # type: ignore
-    except Exception:  # repo-root context
-        from mcp_servers.google_tasks.auth import get_tasks_service  # type: ignore
-
 from .clear_completed_tasks import clear_completed_tasks
 from .create_task import create_task
 from .create_tasklist import create_tasklist
@@ -28,7 +14,6 @@ from .update_task import update_task
 from .update_tasklist import update_tasklist
 
 __all__ = [
-    "tasks_service_context",
     "clear_completed_tasks",
     "create_task",
     "create_tasklist",
@@ -44,18 +29,3 @@ __all__ = [
     "update_task",
     "update_tasklist",
 ]
-
-# Context variable to allow per-request override if needed.
-# We lazily create the Google Tasks service on first access after environment
-# variables are guaranteed to be loaded.
-
-tasks_service_context: contextvars.ContextVar = contextvars.ContextVar(
-    "google_tasks_service", default=None
-)
-
-def get_service():
-    svc = tasks_service_context.get()
-    if svc is None:
-        svc = get_tasks_service()
-        tasks_service_context.set(svc)
-    return svc
