+++ mcp_servers/google_tasks/tools/move_task.py
@@ -5,7 +5,13 @@ from typing import Any, Dict, Optional
 
 from googleapiclient.errors import HttpError
 
-from . import get_service
+try:  # package context
+    from ..auth import get_service  # type: ignore
+except Exception:
+    try:  # script context
+        from auth import get_service  # type: ignore
+    except Exception:  # repo-root context
+        from mcp_servers.google_tasks.auth import get_service  # type: ignore
 
 logger = logging.getLogger(__name__)
